Adds struct literals, member access, variable sized registers.

+++ src/main/java/com/github/ahhoefel/ir/Representation.java
@@ -36,7 +36,7 @@ public class Representation {
     return operations.size();
   }
 
-  public Operation getOperation(int i) {
-    return operations.get(i);
+  public Operation getOperation(long i) {
+    return operations.get((int) i);
   }
 }
